+++ Strikers1945/BackgroundManager.h
@@ -20,7 +20,6 @@ public:
 	void Release();
 	void Render(HDC hdc);
 	void Update();
-	int GetCurrPosY();
-
+	float GetCurrPosY();
 };
